Remove initial legacy-mode implementation in favour of stream-syntax
It doesn't have feature parity yet, but this is the direction
we're going to go in
