--- legacy-modes/style/codemirror.css
@@ -1,23 +1,0 @@
-/* DEFAULT THEME */
-
-.cm-keyword {color: #708;}
-.cm-atom {color: #219;}
-.cm-number {color: #164;}
-.cm-def {color: #00f;}
-.cm-variable,
-.cm-punctuation,
-.cm-property,
-.cm-operator {}
-.cm-variable-2 {color: #05a;}
-.cm-variable-3, .cm-type {color: #085;}
-.cm-comment {color: #a50;}
-.cm-string {color: #a11;}
-.cm-string-2 {color: #f50;}
-.cm-meta {color: #555;}
-.cm-qualifier {color: #555;}
-.cm-builtin {color: #30a;}
-.cm-bracket {color: #997;}
-.cm-tag {color: #170;}
-.cm-attribute {color: #00c;}
-.cm-hr {color: #999;}
-.cm-link {color: #00c;}
